+++ srcs/exec/exec_pre_in_post.c
@@ -6,7 +6,7 @@
 /*   By: jjaniec <jjaniec@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/06/26 10:30:52 by sbrucker          #+#    #+#             */
-/*   Updated: 2018/11/09 17:56:00 by sbrucker         ###   ########.fr       */
+/*   Updated: 2018/11/11 16:44:26 by sbrucker         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -61,63 +61,6 @@ static void	exec_node(char **argv, t_exec *exe, t_ast *node)
 			NULL, exe, node);
 	if (not)
 		exe->ret = (exe->ret == 0) ? 1 : 0;
-}
-
-/*
-** Determines if string should be passed to handle_quotes_expansions
-** to remove quotes / backslashes / replace expansion specifiers
-*/
-
-static int	is_clean_needed_expansions(char **data)
-{
-	char	*ptr;
-
-	ptr = EXPANSIONS_SPECIFIERS;
-	while (*ptr)
-		if (ft_strchr(*data, *ptr++))
-			return (1);
-	return (0);
-}
-
-
-static int	is_clean_needed(char **data)
-{
-	if (*data && (ft_strchr(*data, '\'') || \
-			ft_strchr(*data, '"') || ft_strchr(*data, '\\') || \
-			is_clean_needed_expansions(data)))
-		return (1);
-	return (0);
-}
-
-static void	remove_empty_data(char **data, int i)
-{
-	free(data[i]);
-	while (data[i])
-	{
-		data[i] = data[i + 1];
-		i++;
-	}
-}
-
-static void	clean_data(char **data)
-{
-	int		i;
-	int		quoted;
-
-	i = 0;
-	while (data && data[i])
-	{
-		quoted = ft_strchr(data[i], '"') || ft_strchr(data[i], '`') \
-		|| ft_strchr(data[i], '\'');
-		if (is_clean_needed(data + i))
-			handle_quotes_expansions(data + i);
-		if (!data[i][0] && !quoted)
-		{
-			remove_empty_data(data, i);
-			i--;
-		}
-		i++;
-	}
 }
 
 /*
