+++ srcs/syntax_highlighting/print_colorized_input.c
@@ -6,7 +6,7 @@
 /*   By: jjaniec <jjaniec@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/08/22 17:38:26 by jjaniec           #+#    #+#             */
-/*   Updated: 2018/10/02 14:25:33 by jjaniec          ###   ########.fr       */
+/*   Updated: 2018/11/11 16:53:32 by sbrucker         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
